+++ byteps/common/global.h
@@ -114,7 +114,7 @@ class BytePSGlobal {
 
   static bool pre_run;
   static std::vector <int> _grad_checkpoint;
-  static std::vector<int> _backward_exec;
+  static std::vector<double> _backward_exec;
   static int total_grad;
 
   static std::shared_ptr<NcclManager> GetNccl() { return _nccl_manager; }
