+++ byteps/common/scheduled_queue.h
@@ -59,8 +59,8 @@ namespace byteps {
             std::multiset <std::shared_ptr<TensorTableEntry>, comparator> _ms;
             std::vector <std::shared_ptr<TensorTableEntry>> _mysq;
             bool finish_tag[1600] = {false};
-            long long _grad_tic[1600] = {0};
-            long long _push_start_tic[1600] = {0};
+            double _grad_tic[1600] = {0};
+            double _push_start_tic[1600] = {0};
             int _tensor_part[1600] = {0};
             int _visited[1600] = {0};
             std::stack<int> _mystack;
